Move bytesForHumans to kolibri.utils
